fix(ios): keep transport alive on peer disconnect
peerDisconnected was calling disconnect() which tore down the
entire transport, wiped all state, and forced a full reconnect
cycle. This caused 30-60s connection drops, signal showing
'No Signal', and conversation load timeouts.

Changes:
- peerDisconnected now sets .reconnecting instead of calling
  disconnect(); the relay auto-reconnects and re-sends sync
- Flush loop syncs connectionQuality.transportState every tick
  so signal bars reflect actual transport state immediately
- transportReconnecting also updates connectionQuality state
- ConnectionQualityView renders during .reconnecting state
- ContentView shows TabListView during .reconnecting (keeps
  tabs visible instead of flashing disconnected screen)
- handleSnapshot re-sends in-flight conversation load requests
  that may have been silently dropped during transport hiccup
- 30s safety timer forces full reconnect if .reconnecting
  lingers (e.g. peer truly gone)
- Extract handleSnapshot to SessionViewModel+Snapshot.swift
  to keep EventHandlers.swift under the 600-line cap
